+++ src/include/simeng/models/emulation/Core.hh
@@ -46,7 +46,7 @@ class Core : public simeng::Core {
 
   /** Change the value of the ARM64_SYSREG_CNTVCT_EL0 system register to number
    * of cycles completed.*/
-  void incCNTVCT(int iterations) override;
+  void incCNTVCT(uint64_t iterations) override;
 
  private:
   /** Execute an instruction. */
